requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.6.0
streamlit>=1.35.0
python-dotenv>=1.0.0